split encoder and builder
